move all `xvd` related modules into a new `msxivc` crate
