+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "3"
-members = ["xodus", "xodus-cli", "xodus-service"]
+members = ["msxivc","xodus", "xodus-cli", "xodus-service"]
 
 [workspace.dependencies]
 chrono = "0.4.42"
