+++ msxivc/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "msxivc"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+tokio = { workspace = true }
+reqwest = { workspace = true }
+aes = "0.9.0"
+async-trait = "0.1.89"
+bitflags = "2.13.0"
+bytes = "1.12.0"
+chrono = { workspace = true }
+futures-util = "0.3.32"
+gpt = "4.1.0"
+ntfs = "0.4.0"
+num_enum = "0.7.6"
+thiserror = "2.0.18"
+tokio-util = { version = "0.7.18", features = ["io-util"] }
+uuid = "1.23.1"
+zerocopy = { version = "0.8.48", features = ["derive"] }
