+++ msxivc/src/lib.rs
@@ -1,5 +1,6 @@
 mod crypt;
 pub mod math;
+pub mod models;
 pub mod streaming;
 pub mod streaming_ntfs;
 pub mod utils;
