+++ msxivc/src/models.rs
