+++ msxivc/src/models/constants.rs
