+++ msxivc/src/models/enums.rs
