+++ msxivc/src/models/flags.rs
