+++ msxivc/src/models/structs/parsed.rs
@@ -1,13 +1,11 @@
 use super::raw;
-use crate::models::xvd::constants::{
-    LEGACY_SECTOR_SIZE, SECTOR_SIZE, XVD_HEADER_INCL_SIGNATURE_SIZE,
-};
-use crate::models::xvd::enums::{XvcRegionId, XvdContentType, XvdType};
-use crate::models::xvd::flags::{
+use crate::math::{bytes_to_pages, calculate_number_of_hash_pages, page_number_to_offset};
+use crate::models::constants::{LEGACY_SECTOR_SIZE, SECTOR_SIZE, XVD_HEADER_INCL_SIGNATURE_SIZE};
+use crate::models::enums::{XvcRegionId, XvdContentType, XvdType};
+use crate::models::flags::{
     WriteablePolicyFlags, XvcInfoFlags, XvcRegionFlags, XvcRegionPresenceInfoFlags,
     XvdSegmentMetadataSegmentFlags, XvdVolumeFlags,
 };
-use crate::xvd::math::{bytes_to_pages, calculate_number_of_hash_pages, page_number_to_offset};
 
 use std::collections::HashMap;
 use std::fmt::{Debug, Display};
