+++ msxivc/src/models/structs/raw.rs
