+++ xodus-cli/src/commands/extract.rs
@@ -1,7 +1,5 @@
-use xodus::{
-    tokens::TokenManager,
-    xvd::utils::{XvdFile, unpack_file},
-};
+use msxivc::utils::{XvdFile, unpack_file};
+use xodus::tokens::TokenManager;
 
 use crate::license::get_license;
 pub async fn run(
@@ -28,6 +26,6 @@ pub async fn run(
     }
     if let Some((_, content_key)) = game_splicense.content_keys.into_iter().next() {
         let unpacked = content_key.unpack(&key).expect("failed to unpack");
-        unpack_file(xvd, path.to_string(), destination.to_string(), unpacked).expect("unpack ok");
+        unpack_file(xvd, path.to_string(), destination.to_string(), *unpacked).expect("unpack ok");
     }
 }
