+++ xodus/Cargo.toml
@@ -10,7 +10,6 @@ reqwest = { workspace = true }
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
 quick-xml = { version = "0.40.1", features = ["serialize", "overlapped-lists"] }
-async-trait = "0.1.89"
 log = "0.4.28"
 chrono = { workspace = true }
 base64 = "0.22.1"
@@ -30,12 +29,6 @@ num_enum = "0.7.6"
 prost = "0.14.4"
 globset = "0.4.18"
 thiserror = "2.0.18"
-ntfs = "0.4.0"
-gpt = "4.1.0"
-bitflags = "2.13.0"
-futures-util = "0.3.32"
-bytes = "1.12.0"
-tokio-util = { version = "0.7.18", features = ["io-util"] }
 
 [target.'cfg(not(target_os = "linux"))'.dependencies]
 smbios-lib = "0.9.2"
