+++ xodus/src/lib.rs
@@ -6,7 +6,6 @@ pub mod licensing;
 pub mod models;
 pub mod secrets;
 pub mod tokens;
-pub mod xvd;
 
 pub const XBOX_LIVE_PACKAGES_PC: &str = "https://packagespc.xboxlive.com";
