+++ xodus/src/models/mod.rs
@@ -9,4 +9,3 @@ pub mod soap;
 pub mod xbox;
 pub mod xgameruntime;
 pub mod xodus;
-pub mod xvd;
